Add a meta::random namespace; reorganize rng-related things.
